value_の型をlongへ変更

+++ include/base_node.h
@@ -11,11 +11,11 @@ public:
 
     virtual unsigned int ChildSize() const = 0;
     virtual unsigned int Count() const = 0;
-    virtual int Value() const = 0;
+    virtual long Value() const = 0;
 
     virtual bool IsLeaf() const = 0;
     virtual void Expand(const BaseVrp& vrp, const Solution& solution) = 0;
-    virtual void Update(int value) = 0;
+    virtual void Update(long value) = 0;
 };
 
 #endif /* VRPSOLVER_CPP_BASE_NODE_H */
